音频、视频

+++ src/cgc/js/unit.js
@@ -850,12 +850,20 @@ var vm = {
 
   getTagContent: function(e) {
     e.stopPropagation();
+
     $target = $(e.target);
     vm.data.tagListIndex = $target.index();
 
-    ApiBridge.callNative('ClientVideoManager', 'deleteById', {
-      mediaid: vm.data.mediaid,
-    });
+    if (vm.data.tagListIndex + 1 == 4) {
+      ApiBridge.callNative('ClientVideoManager', 'deleteById', {
+        mediaid: vm.data.mediaid,
+      });
+    }
+    if (vm.data.tagListIndex + 1 == 3) {
+      ApiBridge.callNative('ClientAudioManager', 'deleteById', {
+        mediaid: vm.data.mediaid,
+      });
+    }
 
     if ($('.c-tab-bd ul').eq($target.index()).html() == '') {
       vm.tagList($target.index());
@@ -1414,49 +1422,49 @@ var vm = {
       var html = '';
 
       /*
-        data.map(function(v) {
-          if (v['mediatype'] == 4) {
-            html += '<li newsid=' + v['newsid'] + ' mediatype=' + v['mediatype'] + ' userId=' + v['userid'] + ' class=media-audio>' + '<a class="c-att-t" userid=' + v['userid'] + ' username=' + v['username'] + ' userpic=' + v['userpic'] + ' usertime=' + (v['publishtime'] || '') + ' usertitle=' + v['title'] + ' userdesc=' + v['description'] + '>' + (v['isattention'] == 1 ? '已关注' : '+ 关注') + '</a>' + '<div class=c-media-info><img userId=' + v['userid'] + ' class="c-auth-img" src=' + v['userpic'] + ' alt="">' + '<p userId=' + v['userid'] + ' class="c-auth-title">' + v['username'] + '</p></div>' + '<div class="c-media-audio">' + '<div mediatype=' + v['mediatype'] + ' title=' + v['title'] + ' thumbnailpics=' + v['thumbnailpics'] + ' playtime=' + v['playtime'] + ' status=' + v['status'] + ' mediaid=' + v['mediaid'] + ' class="c-tag-media">' + ((v['mediatype'] == 3 || v['mediatype'] == 4) ? '<span class="c-tag-video"></span>' : '') + '<img class="c-auth-info-img" src=' + v['indexdetail'] + ' alt=""></div><span class="c-tab-jj ">' + ((v['mediatype'] == 1 || v['mediatype'] == 4 || v['mediatype'] == 3) ? v['title'] : v['description']) + '</span></div>' + '<p class="span c-tab-ue">' + '<span class="c-zan"><span class="zan-icon"></span><span class="c-num">' + v['praisenum'] + '</span></span>' + '<span class="c-common" newsid=' + v['newsid'] + ' type=' + v['mediatype'] + '><span class="c-num">' + v['replycount'] + '</span></span>' + '</p>' + '<span class="c-looked">' + v['pv'] + ' 浏览</span>' + '</li>';
-          } else if (v['mediatype'] == 2) {
+      data.map(function(v) {
+        if (v['mediatype'] == 4) {
+          html += '<li newsid=' + v['newsid'] + ' mediatype=' + v['mediatype'] + ' userId=' + v['userid'] + ' class=media-audio>' + '<a class="c-att-t" userid=' + v['userid'] + ' username=' + v['username'] + ' userpic=' + v['userpic'] + ' usertime=' + (v['publishtime'] || '') + ' usertitle=' + v['title'] + ' userdesc=' + v['description'] + '>' + (v['isattention'] == 1 ? '已关注' : '+ 关注') + '</a>' + '<div class=c-media-info><img userId=' + v['userid'] + ' class="c-auth-img" src=' + v['userpic'] + ' alt="">' + '<p userId=' + v['userid'] + ' class="c-auth-title">' + v['username'] + '</p></div>' + '<div class="c-media-audio">' + '<div mediatype=' + v['mediatype'] + ' title=' + v['title'] + ' thumbnailpics=' + v['thumbnailpics'] + ' playtime=' + v['playtime'] + ' status=' + v['status'] + ' mediaid=' + v['mediaid'] + ' class="c-tag-media">' + ((v['mediatype'] == 3 || v['mediatype'] == 4) ? '<span class="c-tag-video"></span>' : '') + '<img class="c-auth-info-img" src=' + v['indexdetail'] + ' alt=""></div><span class="c-tab-jj ">' + ((v['mediatype'] == 1 || v['mediatype'] == 4 || v['mediatype'] == 3) ? v['title'] : v['description']) + '</span></div>' + '<p class="span c-tab-ue">' + '<span class="c-zan"><span class="zan-icon"></span><span class="c-num">' + v['praisenum'] + '</span></span>' + '<span class="c-common" newsid=' + v['newsid'] + ' type=' + v['mediatype'] + '><span class="c-num">' + v['replycount'] + '</span></span>' + '</p>' + '<span class="c-looked">' + v['pv'] + ' 浏览</span>' + '</li>';
+        } else if (v['mediatype'] == 2) {
 
-            var qingImg = '<div class="c-qing-img-wp">';
-            v['indexdetail'].map(function(k, i) {
-              if (i > 2) {
-                return;
-              }
-              qingImg += '<img class="c-qing-img" src=' + k + ' />'
-            })
-            if (v['indexdetail'].length > 3) {
-              qingImg += '<span class="c-qing-num">' + v['indexdetail'].length + '</span></div>'
+          var qingImg = '<div class="c-qing-img-wp">';
+          v['indexdetail'].map(function(k, i) {
+            if (i > 2) {
+              return;
             }
-            html += '<li newsid=' + v['newsid'] + ' mediatype=' + v['mediatype'] + ' userId=' + v['userid'] + ' class=media-qing>' + '<a class="c-att-t ' + (v['isattention'] == '1' ? 'on' : '') + '" userid=' + v['userid'] + ' username=' + v['username'] + ' userpic=' + v['userpic'] + ' usertime=' + (v['publishtime'] || '') + ' usertitle=' + v['title'] + ' userdesc=' + v['description'] + '>' + (v['isattention'] ? '已关注' : '+ 关注') + '</a>' + '<div class=c-media-info><img userId=' + v['userid'] + ' class="c-auth-img" src=' + v['userpic'] + ' alt="">' + '<p userId=' + v['userid'] + ' class="c-auth-title">' + v['username'] + '</p></div>' + '<div class="c-media-audio">' + '<div mediatype=' + v['mediatype'] + ' title=' + v['title'] + ' thumbnailpics=' + v['thumbnailpics'] + ' playtime=' + v['playtime'] + ' status=' + v['status'] + ' mediaid=' + v['mediaid'] + ' class="c-tag-media">' + ((v['mediatype'] == 3 || v['mediatype'] == 4) ? '<span class="c-tag-video"></span>' : '') +  '</div><span class="c-tab-jj ">' + ((v['mediatype'] == 3 || v['mediatype'] == 4 || v['mediatype'] == 1) ? v['title'] : v['description']) + '</span>'+qingImg +'</div>' + '<p class="span c-tab-ue">' + '<span class="c-zan"><span class="zan-icon"></span><span class="c-num">' + v['praisenum'] + '</span></span>' + '<span class="c-common" newsid=' + v['newsid'] + ' type=' + v['mediatype'] + '><span class="c-num">' + v['replycount'] + '</span></span>' + '</p>' + '<span class="c-looked">' + v['pv'] + ' 浏览</span>' + '</li>';
-          } else {
-
-            html += '<li newsid=' + v['newsid'] + ' mediatype=' + v['mediatype'] + ' userId=' + v['userid'] + ' >' + '<a class="c-att-t" userid=' + v['userid'] + ' username=' + v['username'] + ' userpic=' + v['userpic'] + ' usertime=' + (v['publishtime'] || '') + ' usertitle=' + v['title'] + ' userdesc=' + v['description'] + ' >' + (v['isattention'] == 1 ? '已关注' : '+ 关注') + '</a>' + '<div class=c-media-info><img userId=' + v['userid'] + ' class="c-auth-img" src=' + v['userpic'] + ' alt="">' + '<p userId=' + v['userid'] + ' class="c-auth-title">' + v['username'] + '</p></div>' + '<p class="c-tab-jj ' + (v['mediatype'] == 1 ? 'short' : 'long') + '">' + ((v['mediatype'] == 1 || v['mediatype'] == 4 || v['mediatype'] == 3) ? v['title'] : v['description']) + '</p>' + '<div mediatype=' + v['mediatype'] + ' title=' + v['title'] + ' thumbnailpics=' + v['thumbnailpics'] + ' playtime=' + v['playtime'] + ' status=' + v['status'] + ' mediaid=' + v['mediaid'] + ' class="c-tag-media">' + ((v['mediatype'] == 3 || v['mediatype'] == 4) ? '<span class="c-tag-video"></span>' : '') + '<img class="c-auth-info-img" src=' + v['indexdetail'] + ' alt=""></div>' + '<p class="span c-tab-ue">' + '<span class="c-zan"><span class="zan-icon"></span><span class="c-num">' + v['praisenum'] + '</span></span>' + '<span class="c-common" newsid=' + v['newsid'] + ' type=' + v['mediatype'] + '><span class="c-num">' + v['replycount'] + '</span></span>' + '</p>' + '<span class="c-looked">' + v['pv'] + ' 浏览</span>' + '</li>';
+            qingImg += '<img class="c-qing-img" src=' + k + ' />'
+          })
+          if (v['indexdetail'].length > 3) {
+            qingImg += '<span class="c-qing-num">' + v['indexdetail'].length + '</span></div>'
           }
-        })
-
-        if (!vm.data.isLoad) {
-          $('.c-tab-bd ul').eq(index).append(html);
+          html += '<li newsid=' + v['newsid'] + ' mediatype=' + v['mediatype'] + ' userId=' + v['userid'] + ' class=media-qing>' + '<a class="c-att-t ' + (v['isattention'] == '1' ? 'on' : '') + '" userid=' + v['userid'] + ' username=' + v['username'] + ' userpic=' + v['userpic'] + ' usertime=' + (v['publishtime'] || '') + ' usertitle=' + v['title'] + ' userdesc=' + v['description'] + '>' + (v['isattention'] ? '已关注' : '+ 关注') + '</a>' + '<div class=c-media-info><img userId=' + v['userid'] + ' class="c-auth-img" src=' + v['userpic'] + ' alt="">' + '<p userId=' + v['userid'] + ' class="c-auth-title">' + v['username'] + '</p></div>' + '<div class="c-media-audio">' + '<div mediatype=' + v['mediatype'] + ' title=' + v['title'] + ' thumbnailpics=' + v['thumbnailpics'] + ' playtime=' + v['playtime'] + ' status=' + v['status'] + ' mediaid=' + v['mediaid'] + ' class="c-tag-media">' + ((v['mediatype'] == 3 || v['mediatype'] == 4) ? '<span class="c-tag-video"></span>' : '') + '</div><span class="c-tab-jj ">' + ((v['mediatype'] == 3 || v['mediatype'] == 4 || v['mediatype'] == 1) ? v['title'] : v['description']) + '</span>' + qingImg + '</div>' + '<p class="span c-tab-ue">' + '<span class="c-zan"><span class="zan-icon"></span><span class="c-num">' + v['praisenum'] + '</span></span>' + '<span class="c-common" newsid=' + v['newsid'] + ' type=' + v['mediatype'] + '><span class="c-num">' + v['replycount'] + '</span></span>' + '</p>' + '<span class="c-looked">' + v['pv'] + ' 浏览</span>' + '</li>';
         } else {
-          $('.c-tab-bd ul').eq(index).html(html);
+
+          html += '<li newsid=' + v['newsid'] + ' mediatype=' + v['mediatype'] + ' userId=' + v['userid'] + ' >' + '<a class="c-att-t" userid=' + v['userid'] + ' username=' + v['username'] + ' userpic=' + v['userpic'] + ' usertime=' + (v['publishtime'] || '') + ' usertitle=' + v['title'] + ' userdesc=' + v['description'] + ' >' + (v['isattention'] == 1 ? '已关注' : '+ 关注') + '</a>' + '<div class=c-media-info><img userId=' + v['userid'] + ' class="c-auth-img" src=' + v['userpic'] + ' alt="">' + '<p userId=' + v['userid'] + ' class="c-auth-title">' + v['username'] + '</p></div>' + '<p class="c-tab-jj ' + (v['mediatype'] == 1 ? 'short' : 'long') + '">' + ((v['mediatype'] == 1 || v['mediatype'] == 4 || v['mediatype'] == 3) ? v['title'] : v['description']) + '</p>' + '<div mediatype=' + v['mediatype'] + ' title=' + v['title'] + ' thumbnailpics=' + v['thumbnailpics'] + ' playtime=' + v['playtime'] + ' status=' + v['status'] + ' mediaid=' + v['mediaid'] + ' class="c-tag-media">' + ((v['mediatype'] == 3 || v['mediatype'] == 4) ? '<span class="c-tag-video"></span>' : '') + '<img class="c-auth-info-img" src=' + v['indexdetail'] + ' alt=""></div>' + '<p class="span c-tab-ue">' + '<span class="c-zan"><span class="zan-icon"></span><span class="c-num">' + v['praisenum'] + '</span></span>' + '<span class="c-common" newsid=' + v['newsid'] + ' type=' + v['mediatype'] + '><span class="c-num">' + v['replycount'] + '</span></span>' + '</p>' + '<span class="c-looked">' + v['pv'] + ' 浏览</span>' + '</li>';
         }
+      })
 
-        $('.c-loading').hide();
-        vm.data.isLoad = true;
+      if (!vm.data.isLoad) {
+        $('.c-tab-bd ul').eq(index).append(html);
+      } else {
+        $('.c-tab-bd ul').eq(index).html(html);
+      }
 
-        $('.c-tab-bd ul li').each(function(v, i) {
-          if ($(this).attr('mediatype') == 4) {
-            return
-          };
-          $(this).find('.c-auth-info-img').height($(this).find('.c-auth-info-img').width() * 0.5625);
-          if ($(this).attr('mediatype') == 2) {
-            //$(this).find('.c-qing-img').height($(this).find('.c-qing-img').width() * 0.6);
-          }
-        })
+      $('.c-loading').hide();
+      vm.data.isLoad = true;
 
-        return;
-        */
+      $('.c-tab-bd ul li').each(function(v, i) {
+        if ($(this).attr('mediatype') == 4) {
+          return
+        };
+        $(this).find('.c-auth-info-img').height($(this).find('.c-auth-info-img').width() * 0.5625);
+        if ($(this).attr('mediatype') == 2) {
+          //$(this).find('.c-qing-img').height($(this).find('.c-qing-img').width() * 0.6);
+        }
+      })
+
+      return;
+      */
       //本地关注与线上数据判断已关注过滤
       //登录未登录 
 
@@ -1485,9 +1493,9 @@ var vm = {
                 } else if (v['mediatype'] == 2) {
 
                   var qingImg = '<div class="c-qing-img-wp">';
-                  if(v['indexdetail'].length < 3){
+                  if (v['indexdetail'].length < 3) {
                     qingImg = '<div class="c-qing-img-one"><img class="c-qing-img" src=' + v['indexdetail'][0] + ' /></div>'
-                  }else{
+                  } else {
                     v['indexdetail'].map(function(k, i) {
                       if (i > 2) {
                         return;
@@ -1495,7 +1503,7 @@ var vm = {
                       qingImg += '<img class="c-qing-img" src=' + k + ' />'
                     })
                   }
-                  
+
                   if (v['indexdetail'].length > 3) {
                     qingImg += '<span class="c-qing-num">' + v['indexdetail'].length + '</span>'
                   } else {
@@ -1526,8 +1534,7 @@ var vm = {
               $('.c-loading').hide();
               vm.data.isLoad = true;
             })
-          } catch (e) {
-          }
+          } catch (e) {}
         } else {
           data.map(function(v) {
             if (v['mediatype'] == 4) {
@@ -1536,9 +1543,9 @@ var vm = {
 
               var qingImg = '<div class="c-qing-img-wp">';
 
-              if(v['indexdetail'].length < 3){
+              if (v['indexdetail'].length < 3) {
                 qingImg = '<div class="c-qing-img-one"><img class="c-qing-img" src=' + v['indexdetail'][0] + ' /></div>'
-              }else{
+              } else {
                 v['indexdetail'].map(function(k, i) {
                   if (i > 2) {
                     return;
@@ -1636,14 +1643,17 @@ if (/tag-name/.test(window.location.href)) {
     if (!!vm.data.isLoad) {
       vm.data.isLoad = false;
       $('.c-loading').show();
-      //删除视频 
-      ApiBridge.callNative('ClientVideoManager', 'deleteById', {
-        mediaid: vm.data.mediaid,
-      });
-      //删除音频 
-      ApiBridge.callNative('ClientAudioManager', 'deleteById', {
-        mediaid: vm.data.mediaid,
-      });
+
+      if (vm.data.tagListIndex + 1 == 4) {
+        ApiBridge.callNative('ClientVideoManager', 'deleteById', {
+          mediaid: vm.data.mediaid,
+        });
+      }
+      if (vm.data.tagListIndex + 1 == 3) {
+        ApiBridge.callNative('ClientAudioManager', 'deleteById', {
+          mediaid: vm.data.mediaid,
+        });
+      }
       vm.tagList(vm.data.tagListIndex);
     }
   });
@@ -1659,10 +1669,17 @@ if (/tag-name/.test(window.location.href)) {
     container: '.container',
     beforePull: function() {
       console.log('beforePull')
-        //删除视频 
-      ApiBridge.callNative('ClientVideoManager', 'deleteById', {
-        mediaid: vm.data.mediaid,
-      });
+
+      if (vm.data.tagListIndex + 1 == 4) {
+        ApiBridge.callNative('ClientVideoManager', 'deleteById', {
+          mediaid: vm.data.mediaid,
+        });
+      }
+      if (vm.data.tagListIndex + 1 == 3) {
+        ApiBridge.callNative('ClientAudioManager', 'deleteById', {
+          mediaid: vm.data.mediaid,
+        });
+      }
     },
     onRefresh: function() {
       vm.tagList(vm.data.tagListIndex);
